combine tokenizing and expanding
