+++ Makefile
@@ -8,11 +8,12 @@ INCLUDES = -I./includes/ -I./libft/includes/
 SRCS = main.c \
 	./tokenizing/tokenizing.c ./tokenizing/clean_tokens.c \
 	./tokenizing/print_tokens.c ./tokenizing/tokenizing_utils.c \
+	./tokenizing/handle_pipe_redir.c ./tokenizing/handle_quotes.c \
+	./tokenizing/handle_variable.c ./tokenizing/handle_word.c \
 	./parsing/parsing.c ./parsing/print_ast.c ./parsing/addArg.c \
 	./parsing/addRedir.c ./parsing/cleanAST.c ./parsing/createNode.c \
 	./parsing/isRedirection.c ./parsing/parseCommand.c \
 	./parsing/parseExpression.c ./parsing/parseRedirection.c \
-	./expanding/expanding.c ./expanding/print_exps.c
 
 OBJS = $(SRCS:.c=.o)
