+++ includes/minishell.h
@@ -6,12 +6,18 @@
 /*   By: yingzhan <yingzhan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/19 17:03:16 by yingzhan          #+#    #+#             */
-/*   Updated: 2025/09/02 16:04:10 by yingzhan         ###   ########.fr       */
+/*   Updated: 2025/09/05 12:03:58 by yingzhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
+
+typedef struct s_shell
+{
+	char	**env;
+	int		exit_code;
+}	t_shell;
 
 #include <stdio.h>
 #include <readline/readline.h>
@@ -21,8 +27,5 @@
 #include "../libft/includes/get_next_line.h"
 #include "parsing.h"
 #include "tokenizing.h"
-#include "expanding.h"
-
-
 
 #endif
