+++ tokenizing/handle_quotes.c
@@ -1,27 +1,36 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   expanding.h                                        :+:      :+:    :+:   */
+/*   handle_quotes.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yingzhan <yingzhan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/09/02 16:14:06 by yingzhan          #+#    #+#             */
-/*   Updated: 2025/09/04 12:03:11 by yingzhan         ###   ########.fr       */
+/*   Created: 2025/09/05 18:51:25 by yingzhan          #+#    #+#             */
+/*   Updated: 2025/09/05 19:50:05 by yingzhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef EXPANDING_H
-# define EXPANDING_H
+#include "../includes/minishell.h"
 
-# include "parsing.h"
-
-typedef struct s_shell
+int	change_quote(char c, int *quote)
 {
-	char	**env;
-	int		exit_code;
-}	t_shell;
-
-void	print_exp(t_node *node);
-void	ft_expand(t_node *root, t_shell *shell);
-
-#endif
+	if (c == '\'')
+	{
+		if (*quote == q_sopen)
+			*quote = q_close;
+		else if (*quote == q_close)
+			*quote = q_sopen;
+		else
+			return (0);
+	}
+	else if (c == '"')
+	{
+		if (*quote == q_dopen)
+			*quote = q_close;
+		else if (*quote == q_close)
+			*quote = q_dopen;
+		else
+			return (0);
+	}
+	return (1);
+}
